upload results

+++ tpcc.cc
@@ -168,19 +168,19 @@ void welcome(int argc, const char *const *argv) {
 }
 
 void tableSize(TPCCTables *tables, bool is_initial, bool is_compressed) {
-    int ini_warehouses = tables->warehouseSize(num_warehouses);
-    int ini_districts = tables->districtSize(num_warehouses);
-    int ini_customers = tables->customerSize(num_warehouses);
-    int ini_orders = tables->orderSize(num_warehouses, 0);
-    int ini_orderline = tables->orderlineSize(num_warehouses, 0);
-    int ini_neworders = tables->newOrderSize();
-    int ini_items = tables->itemSize();
-    int ini_stocks = tables->stockSize(num_warehouses);
-    int ini_history = tables->historySize();
+    int64_t ini_warehouses = tables->warehouseSize(num_warehouses);
+    int64_t ini_districts = tables->districtSize(num_warehouses);
+    int64_t ini_customers = tables->customerSize(num_warehouses);
+    int64_t ini_orders = tables->orderSize(num_warehouses, 0);
+    int64_t ini_orderline = tables->orderlineSize(num_warehouses, 0);
+    int64_t ini_neworders = tables->newOrderSize();
+    int64_t ini_items = tables->itemSize();
+    int64_t ini_stocks = tables->stockSize(num_warehouses);
+    int64_t ini_history = tables->historySize();
 
-    int com_customers = tables->customerBlitzSize(num_warehouses);
-    int com_stock = tables->stockBlitzSize(num_warehouses);
-    int com_orderline = tables->orderlineBlitzSize(num_warehouses, NUM_TRANSACTIONS);
+    int64_t com_customers = tables->customerBlitzSize(num_warehouses);
+    int64_t com_stock = tables->stockBlitzSize(num_warehouses);
+    int64_t com_orderline = tables->orderlineBlitzSize(num_warehouses, NUM_TRANSACTIONS);
 
     if (is_initial && !is_compressed) {
         std::cout << "------------ Initial and uncompressed Size ------------ \n";
@@ -193,7 +193,7 @@ void tableSize(TPCCTables *tables, bool is_initial, bool is_compressed) {
         std::cout << "Item: " << ini_items << " byte" << std::endl;
         std::cout << "Stock: " << ini_stocks << " byte" << std::endl;
         std::cout << "History: " << ini_history << " byte" << std::endl;
-        int32_t total =
+        int64_t total =
                 ini_warehouses + ini_districts + ini_customers + ini_orders + ini_orderline +
                 ini_neworders + ini_items + ini_stocks + ini_history;
         std::cout << "Total: " << total << " byte" << std::endl;
@@ -222,7 +222,7 @@ void tableSize(TPCCTables *tables, bool is_initial, bool is_compressed) {
         std::cout << "Item: " << ini_items << " byte" << std::endl;
         std::cout << "Stock: " << com_stock << " byte" << std::endl;
         std::cout << "History: " << ini_history << " byte" << std::endl;
-        int32_t total =
+        int64_t total =
                 ini_warehouses + ini_districts + com_customers + ini_orders + com_orderline +
                 ini_neworders + ini_items + com_stock + ini_history;
         std::cout << "Total: " << total << " byte" << std::endl;
